+++ test/_env/srv/handlers/guest-billing.js
@@ -1,0 +1,5 @@
+"use strict";
+
+const FioriService = require("./fiori");
+
+module.exports = class GuestBillingService extends FioriService {};