+++ test/ws/redis_ws.test.js
@@ -82,7 +82,8 @@ describe("Redis", () => {
     expect(redis.client.subscribe).toHaveBeenNthCalledWith(4, "websocket/cloudevents", expect.any(Function));
     expect(redis.client.subscribe).toHaveBeenNthCalledWith(5, "websocket/fiori", expect.any(Function));
     expect(redis.client.subscribe).toHaveBeenNthCalledWith(6, "websocket/fns-websocket", expect.any(Function));
-    expect(redis.client.subscribe).toHaveBeenNthCalledWith(7, "websocket/main", expect.any(Function));
+    expect(redis.client.subscribe).toHaveBeenNthCalledWith(7, "websocket/guest-billing", expect.any(Function));
+    expect(redis.client.subscribe).toHaveBeenNthCalledWith(8, "websocket/main", expect.any(Function));
     expect(redis.client.publish).toHaveBeenCalledWith(
       "websocket/chat",
       '{"tenant":"t1","event":"received","data":{"text":"test","user":"alice"},"headers":{"header":"value"}}',